建立Order總覽 Part.7

+++ PurchaseOrder/Models/BoxTypeMetadata.cs
@@ -3,14 +3,9 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-using System;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-
 namespace PurchaseOrderSys.Models
 {
-
+    
     /// <summary>
     /// BoxType class
     /// </summary>
@@ -70,7 +65,7 @@ namespace PurchaseOrderSys.Models
     		/// </summary>        
     	    [Display(Name = "BoxType_Weight", ResourceType = typeof(App_GlobalResources.Resource))]
             [Required(ErrorMessageResourceName = "Required", ErrorMessageResourceType = typeof(App_GlobalResources.Resource))]
-    		public double  Weight { get; set; }
+    		public int  Weight { get; set; }
     
     		    
     		/// <summary>
